added a loading "while" loop
Made it so there's loading text, and the dots from an ellipsis move. If it has an error, then it stops all of it and shuts it down. Every time it loads it stops the code for half a second for about 20 loops. The code often replaces the previous text in the file to make it so there's less lines for the users to read.